requirements-3.8.txt
pydantic
python-dotenv
smart_open
ruamel.yaml
pandas